♻️ Delete site, move to xeo-scrum
