--- apps/site-e2e/project.json
@@ -1,28 +1,0 @@
-{
-  "root": "apps/site-e2e",
-  "sourceRoot": "apps/site-e2e/src",
-  "projectType": "application",
-  "targets": {
-    "e2e": {
-      "executor": "@nrwl/cypress:cypress",
-      "options": {
-        "cypressConfig": "apps/site-e2e/cypress.json",
-        "devServerTarget": "site:serve"
-      },
-      "configurations": {
-        "production": {
-          "devServerTarget": "site:serve:production"
-        }
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["apps/site-e2e/**/*.{js,ts}"]
-      }
-    }
-  },
-  "tags": [],
-  "implicitDependencies": ["site"]
-}
