--- apps/site/pages/styles.css
@@ -1,11 +1,0 @@
-@tailwind components;
-@tailwind base;
-@tailwind utilities;
-
-@layer base {
-  html,
-  body {
-    @apply dark:bg-dark-900;
-    @apply accent-primary-300;
-  }
-}
