--- apps/site/project.json
@@ -1,56 +1,0 @@
-{
-  "root": "apps/site",
-  "sourceRoot": "apps/site",
-  "projectType": "application",
-  "port": "4201",
-  "targets": {
-    "build": {
-      "executor": "@nrwl/next:build",
-      "outputs": ["{options.outputPath}"],
-      "defaultConfiguration": "production",
-      "options": {
-        "root": "apps/site",
-        "outputPath": "dist/apps/site"
-      },
-      "configurations": {
-        "production": {}
-      }
-    },
-    "serve": {
-      "executor": "@nrwl/next:server",
-      "options": {
-        "buildTarget": "site:build",
-        "dev": true,
-        "port": 4201
-      },
-      "configurations": {
-        "production": {
-          "buildTarget": "site:build:production",
-          "dev": false
-        }
-      }
-    },
-    "export": {
-      "executor": "@nrwl/next:export",
-      "options": {
-        "buildTarget": "site:build:production"
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/apps/site"],
-      "options": {
-        "jestConfig": "apps/site/jest.config.js",
-        "passWithNoTests": true
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["apps/site/**/*.{ts,tsx,js,jsx}"]
-      }
-    }
-  },
-  "tags": []
-}
