+++ apps/xeo-scrum/components/Background/Background.tsx
