+++ apps/xeo-scrum/components/Footer/Footer.tsx
