+++ apps/xeo-scrum/public/xeo-bdc-dark.png
