+++ apps/xeo-scrum/public/xeo-bdc-light.png
