+++ apps/xeo-scrum/public/xeo-editor-dark.png
