+++ apps/xeo-scrum/public/xeo-editor-light.png
